pyproject.toml
[tool.poetry]
name = "write-offs-api"
version = "0.1.0"
description = ""
authors = ["eldos <eldos.baktybekov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.80.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
sse-starlette = "^1.1.6"
redis = "^4.3.4"
dodo-lib-usbtypec1 = "^1.1.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "src",
]
